Adds tf_backend module
The tf_backend module creates a storage bucket for usage as Terraform's
backend.

It's always good to save Terraform state in somewhere safe,
i.e. not on your laptop. Before we do anything with Terraform, we create
the storage bucket, then use this bucket as the backend (defined on
backend.tf on the next commit).

The test is also interesting: we don't destroy things, we just ensure
they exist.

(project has already been created by other person, so we don't manage it
with Terraform)